Move CLI user creation from impl level to service level
